Take postcode as input, output next arrivals at closest 2 bus stops

+++ BusBoardScratch/LatLong.cs
@@ -2,8 +2,8 @@
 {
     public class LatLong
     {
-        public decimal latitude { get; set; }
-        public decimal longitude { get; set; }
+        public float latitude { get; set; }
+        public float longitude { get; set; }
 
         public override string ToString()
         {
